server/pyproject.toml
[tool.poetry]
name = "at"
version = "0.0.1"
description = "at"
authors = ["at <at@marusoftware.net>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10.0"
fastapi = "^0.115.3"
tortoise-orm = "^0.21.6"
pydantic = {extras = ["email"], version = "^2.9.2"}
pydantic-settings = "^2.5.2"
itsdangerous = "^2.2.0"
aerich = "^0.7.2"
uvicorn = {extras = ["standard"], version = "^0.31.0"}
passlib = "^1.7.4"
python-multipart = "^0.0.18"
aiosmtpd = "^1.4.6"
fastapi-sso = "^0.18.0"

[tool.poetry.group.dev.dependencies]
fastapi = {extras = ["dev"], version = "^0.115.3"}
watchfiles = "^0.24.0"
tortoise-cli = "^0.1.2"


[tool.poetry.group.postgres.dependencies]
tortoise-orm = {extras = ["asyncpg"], version = "^0.21.6"}


[tool.aerich]
tortoise_orm = "at.db.DB_CONFIG"
location = "./migrations"
src_folder = "./."
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
